big update repo

+++ src/routes/products.router.js
@@ -1,6 +1,8 @@
 import { Router } from "express";
 import ProductManager from "../DAOs/ProductManagerMongo.class.js";
 import socketServer from "../App.js";
+import { rolesMiddlewareAdmin } from "../middleware/role.middleware.js";
+import passport from "passport";
 
 const router = Router();
 const ProductsManager = new ProductManager();
@@ -30,12 +32,22 @@ router.get("/stock", async (req, res) => {
   res.send(cars);
 });
 
-router.post("/", async (req, res) => {
-  const car = await ProductsManager.addCar(req.body);
-  socketServer.emit('addCar', car);
-  res.send(car);
-});
-
+router.post(
+  "/",
+  passport.authenticate("jwt", { session: false }),
+  rolesMiddlewareAdmin,
+  async (req, res, next) => {
+    if (req.session.user.role == "admin") {
+      try {
+        const car = await ProductsManager.addCar(req.body);
+        socketServer.emit('addCar', car);
+        res.send(car);
+      } catch (error) {
+        return next(error)
+      }
+    }
+  }
+);
 
 router.get("/:id", async (req, res) => {
   const car = await ProductsManager.getCarById(
@@ -46,21 +58,34 @@ router.get("/:id", async (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
-  const car = await ProductsManager.updateCarById(
-    req.params.id,
-    req.body
-  );
-  if (car.modifiedCount > 0) {
-    const newCar = await ProductsManager.getCarById(req.params.id)
-    socketServer.emit('updateCar', newCar);
+  if (req.session.user.role == "admin") {
+    const car = await ProductsManager.updateCarById(
+      req.params.id,
+      req.body
+    );
+    if (car.modifiedCount > 0) {
+      const newCar = await ProductsManager.getCarById(req.params.id)
+      socketServer.emit('updateCar', newCar);
+    }
+    res.send(car);
+  } else {
+    res.send({ error: "acceso denegado" })
   }
-  res.send(car);
 });
 
-router.delete("/:id", async (req, res) => {
-  const car = await ProductsManager.deleteCar(req.params.id);
-  socketServer.emit('deleteCar', req.params.id);
-  res.send(car);
-});
+router.delete(
+  "/:id",
+  passport.authenticate("jwt", { session: false }),
+  rolesMiddlewareAdmin,
+  async (req, res) => {
+    if (req.session.user.role == "admin") {
+      const car = await ProductsManager.deleteCar(req.params.id);
+      socketServer.emit('deleteCar', req.params.id);
+      res.send(car);
+    } else {
+      res.send({ error: "acceso denegado" })
+    }
+  }
+);
 
 export default router;
